Software Quality Assurance Plan in Software Development
A software quality assurance plan's main goal is to guarantee that the market's product or service is trouble- and bug-free. Additionally, it must fulfill the specifications listed in the SRS (software requirement specification).
An SQA plan serves three purposes. It includes the following:

Determining the QA duties assigned to the concerned team.
A list of the areas that require review, audit, and examination.
Determines the work products for SQA.