Scrum of Agile Software Development - Software Engineering
Scrum is an Iterative and Incremental Development (IID) method that emphasizes a set of project management values and practices, rather than those in requirements, implementation, and so on. As such, it is easily combined with or complementary to other methods.
A key Scrum theme is its emphasis on empirical rather than defined process. Scrum appears simple, yet has practices that deeply influence the work experience and that capture key adaptive and agile qualities.